+++ src/main/java/com/nova/narrativa/domain/llm/controller/StoryController.java
@@ -9,6 +9,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Map;
+
 @RestController
 @RequestMapping("/generate-story")
 public class StoryController {
@@ -29,18 +31,9 @@ public class StoryController {
                     request.getUserId()
             );
             return ResponseEntity.ok(storyResponse);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("잘못된 요청입니다: " + e.getMessage());
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.CONFLICT)
-                    .body("현재 상태로 인해 요청을 처리할 수 없습니다: " + e.getMessage());
-        } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("서버 내부 오류가 발생했습니다: " + e.getMessage());
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("알 수 없는 오류가 발생했습니다: " + e.getMessage());
+                    .body("Error: " + e.getMessage());
         }
     }
 
@@ -51,21 +44,12 @@ public class StoryController {
             String storyResponse = storyService.continueStory(
                     request.getGameId().toString(),  // story_id
                     request.getGenre(),
-                    request.getUserChoice()  // user_choice
+                    request.getUserSelect()  // user_choice
             );
             return ResponseEntity.ok(storyResponse);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("잘못된 요청입니다: " + e.getMessage());
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.CONFLICT)
-                    .body("현재 상태로 인해 요청을 처리할 수 없습니다: " + e.getMessage());
-        } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("서버 내부 오류가 발생했습니다: " + e.getMessage());
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("알 수 없는 오류가 발생했습니다: " + e.getMessage());
+                    .body("Error: " + e.getMessage());
         }
     }
 }
