+++ src/main/java/com/nova/narrativa/domain/llm/dto/ChatRequest.java
@@ -2,13 +2,9 @@ package com.nova.narrativa.domain.llm.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import lombok.*;
+import lombok.Data;
 
 @Data
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
 public class ChatRequest {
 
     @NotNull(message = "게임 ID는 필수입니다.")
@@ -18,7 +14,5 @@ public class ChatRequest {
     private String genre;
 
     @NotEmpty(message = "사용자 선택은 필수입니다.")
-    private String userChoice;
-
-    private int currentStage;
+    private String userSelect;
 }
