+++ src/main/java/com/nova/narrativa/domain/llm/entity/Stage.java
@@ -6,17 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
-public class Story {
+public class Stage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long storyId;
+    private Long stageId;
 
     @ManyToOne
     @JoinColumn(name = "game_id")
@@ -29,15 +27,8 @@ public class Story {
 
     private String userChoice;
 
-    private int choiceCount;
-
     private String imageUrl;
 
     @Column(columnDefinition = "TEXT")
     private String conversationHistory;
-
-    private LocalDateTime startTime;
-
-    private LocalDateTime endTime;
-
 }
